requirements.txt
crewai
openai
instagrapi
moviepy
python-dotenv
groq
schedule
